RDS blueprints (#37)
* Add tests for aurora

* Default parameter group for aurora

Instead of passing NoValue as parameter group when sorting to
default, specify the default explicitely. This is needed for the
new aurora-mysql 5.7 families.

* AuroraMysqlCluster

Add a new type of cluster for aurora-mysql (family for aurora
MySQL 5.7).

* RDS tests

Add tests for rds: ClusterInstance and Postgres.

* Better support for snapshot restore in RDS

Allow to leave master username and password empty when restoring from
snapshot.

Allow not creating an initial database.

* Calm lint

* Add checks for empty user/password in RDS

Add checks for RDS user, password and snapshot restores:
will rise a type error when specifying user or password and
snapshot identifier, or vise versa.

* Remove no-op line

Remove a variable initialisation which is overridden shortly
after.